Math_gest/requirements.txt
opencv-python
temp
streamlit
mediapipe
numpy
Pillow
google-generativeai
speechrecognition
cvzone
streamlit-lottie